pyproject.toml
[project]
name = "vendapay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.5",
    "django-filter>=25.1",
    "djangorestframework>=3.16.1",
    "python-decouple>=3.8",
]
